requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pyee==11.1.1
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.3
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.13.1
zipp==3.20.2
